refactor: split UI into modules
