+++ src/ui/input.rs
@@ -1,0 +1,118 @@
+use super::{
+    app::{App, EditorMode, Tab},
+    editor::Editor,
+};
+use crossterm::event::{KeyCode, KeyEvent, KeyEventKind, KeyModifiers};
+use std::{io, time::Instant};
+
+pub fn handle_key(app: &mut App, key: KeyEvent) -> io::Result<bool> {
+    if key.kind != KeyEventKind::Press {
+        return Ok(false);
+    }
+    let ctrl = key.modifiers.contains(KeyModifiers::CONTROL);
+
+    match app.mode {
+        EditorMode::Insert => {
+            // Special: Esc leaves insert -> command
+            if key.code == KeyCode::Esc {
+                app.mode = EditorMode::Command;
+                return Ok(false);
+            }
+
+            // Assemble (Ctrl+R) também no modo Insert
+            if ctrl && matches!(key.code, KeyCode::Char('r')) {
+                app.assemble_and_load();
+                return Ok(false);
+            }
+
+            match (key.code, app.tab) {
+                // Insert mode: everything types into editor if on Editor tab
+                (code, Tab::Editor) => match code {
+                    KeyCode::Left => app.editor.move_left(),
+                    KeyCode::Right => app.editor.move_right(),
+                    KeyCode::Up => app.editor.move_up(),
+                    KeyCode::Down => app.editor.move_down(),
+                    KeyCode::Backspace => app.editor.backspace(),
+                    KeyCode::Delete => app.editor.delete_char(),
+                    KeyCode::Enter => app.editor.enter(),
+                    KeyCode::Tab => app.editor.insert_spaces(4), // use spaces to avoid cursor width issues
+                    KeyCode::End => {
+                        app.editor.cursor_col = Editor::char_count(app.editor.current_line())
+                    }
+                    KeyCode::Char(c) => app.editor.insert_char(c), // includes '1'/'2'
+                    _ => {}
+                },
+                // In Insert mode, other tabs ignore typing
+                _ => {}
+            }
+            app.editor_dirty = true;
+            app.last_edit_at = Some(Instant::now());
+            app.diag_line = None;
+            app.diag_msg = None;
+            app.diag_line_text = None;
+            app.last_compile_ok = None;
+            app.last_assemble_msg = None;
+        }
+        EditorMode::Command => {
+            // Quit in command mode only
+            if matches!(key.code, KeyCode::Char('q') | KeyCode::Esc) {
+                return Ok(true);
+            }
+
+            // Global assemble
+            if ctrl && matches!(key.code, KeyCode::Char('r')) {
+                app.assemble_and_load();
+                return Ok(false);
+            }
+
+            match (key.code, app.tab) {
+                (KeyCode::Char('i') | KeyCode::Enter, Tab::Editor) => {
+                    app.mode = EditorMode::Insert;
+                    return Ok(false);
+                }
+                // Tab switching only in command mode
+                (KeyCode::Char('1'), _) => app.tab = Tab::Editor,
+                (KeyCode::Char('2'), _) => app.tab = Tab::Run,
+                (KeyCode::Char('3'), _) => app.tab = Tab::Docs,
+
+                // Run controls
+                (KeyCode::Char('s'), Tab::Run) => {
+                    app.single_step();
+                }
+                (KeyCode::Char('r'), Tab::Run) => {
+                    app.is_running = true;
+                }
+                (KeyCode::Char('p'), Tab::Run) => {
+                    app.is_running = false;
+                }
+                (KeyCode::Char('t'), Tab::Run) => {
+                    app.show_registers = !app.show_registers;
+                }
+                (KeyCode::Char('f'), Tab::Run) => {
+                    app.show_hex = !app.show_hex;
+                }
+
+                // Docs scroll
+                (KeyCode::Up, Tab::Docs) => {
+                    app.docs_scroll = app.docs_scroll.saturating_sub(1);
+                }
+                (KeyCode::Down, Tab::Docs) => {
+                    app.docs_scroll += 1;
+                }
+                (KeyCode::PageUp, Tab::Docs) => {
+                    app.docs_scroll = app.docs_scroll.saturating_sub(10);
+                }
+                (KeyCode::PageDown, Tab::Docs) => {
+                    app.docs_scroll += 10;
+                }
+
+                // Editor navigation in command mode (optional)
+                (KeyCode::Up, Tab::Editor) => app.editor.move_up(),
+                (KeyCode::Down, Tab::Editor) => app.editor.move_down(),
+                _ => {}
+            }
+        }
+    }
+
+    Ok(false)
+}
